# Clean up tables code