--- a/src/linear_combination.jl
+++ b/src/linear_combination.jl
@@ -7,16 +7,14 @@ struct LinearCombination <: Transform
     coefficients::Vector{Real}
 end
 
-function _check_dimensions_match(LC::LinearCombination, num_inds)
-    num_coefficients = length(LC.coefficients)
-    if num_inds != num_coefficients
+function _sum_row(row, coefficients)
+    if length(row) != length(coefficients)
         throw(DimensionMismatch(
-            "Size $num_inds doesn't match number of coefficients ($num_coefficients)"
+            "Size $length(row) doesn't match number of coefficients $(length(coefficients))"
         ))
     end
+   return sum(map(*, row, coefficients))
 end
-
-_sum_row(row, coefficients) = sum(map(*, row, coefficients))
 
 """
     apply(
@@ -36,16 +34,12 @@ function apply(
 
     dims === Colon() && throw(ArgumentError("dims=: not supported, choose dims ∈ [1, $N]"))
 
-    # Get the number of slices - error if doesn't match the number of coefficients
-    num_slices = inds === Colon() ? size(A, dims) : length(inds)
-    _check_dimensions_match(LC, num_slices)
-
     return _sum_row(eachslice(selectdim(A, dims, inds); dims=dims), LC.coefficients)
 end
 
 
 """
-    apply(table, LC::LinearCombination; cols=nothing, [header]) -> Table
+    apply(table, LC::LinearCombination; [cols], [header]) -> Table
 
 Applies the [`LinearCombination`](@ref) across the specified cols in `table`. If no `cols`
 are specified, then the [`LinearCombination`](@ref) is applied to all columns.
@@ -53,25 +47,20 @@ are specified, then the [`LinearCombination`](@ref) is applied to all columns.
 Optionally provide a `header` for the output table. If none is provided the default in
 `Tables.table` is used.
 """
-function apply(table, LC::LinearCombination; cols=nothing, kwargs...)
+function apply(table, LC::LinearCombination; cols=_get_cols(table), kwargs...)
     Tables.istable(table) || throw(MethodError(apply, (table, LC)))
 
-    cols = _to_vec(cols)  # handle single column name
-
-    # Error if dimensions don't match
-    num_cols = cols === nothing ? length(Tables.columnnames(table)) : length(cols)
-    _check_dimensions_match(LC, num_cols)
+    # Extract a columns iterator that we should be able to use to mutate the data.
+    # NOTE: Mutation is not guaranteed for all table types, but it avoid copying the data
+    coltable = Tables.columntable(table)
+    cols = _to_vec(cols)
 
     # Keep the generic form when not specifying column names
     # because that is much more performant than selecting each col by name
-    result = if cols === nothing
-        hcat([_sum_row(row, LC.coefficients) for row in Tables.rows(table)])
-    else
-        hcat([
-            _sum_row([row[cname] for cname in cols], LC.coefficients)
-            for row in Tables.rows(table)
-        ])
-    end
+    result = hcat([
+        _sum_row([row[cname] for cname in cols], LC.coefficients)
+        for row in Tables.rows(table)
+    ])
 
     header = get(kwargs, :header, nothing)
     return Tables.materializer(table)(_to_table(result, header))