After #61 I thought it would be good to refactor / clean up as the code was getting a bit messy.

* simplifying the `apply(table, ...)` methods
* removing the 3-arg `apply` methods
* refactoring the linear combination `_sum_row` method to only perform it once for all the cols instead of for every row.

This also comes with some more performance boosts - especially for `LinearCombination` 
```julia
julia> using DataFrames, FeatureTransforms

julia> p = Power(3);

julia> df = DataFrame(rand(100, 100));

# this branch
julia> @time FeatureTransforms.apply(df, p);
  0.542554 seconds (1.74 M allocations: 91.461 MiB, 4.43% gc time)
  0.000439 seconds (1.06 k allocations: 317.812 KiB)

# main
julia> @time FeatureTransforms.apply(df, p);
  0.365457 seconds (1.37 M allocations: 76.569 MiB, 6.15% gc time)
  0.002109 seconds (1.52 k allocations: 4.178 MiB)

# this branch
julia> @time FeatureTransforms.apply(df, p, cols=:x1);
  0.020192 seconds (47.53 k allocations: 2.830 MiB)
  0.000131 seconds (53 allocations: 11.609 KiB)

# main
julia> @time FeatureTransforms.apply(df, p, cols=:x1);
  0.010885 seconds (8.50 k allocations: 513.378 KiB)
  0.000043 seconds (34 allocations: 5.234 KiB)

julia> lc = LinearCombination(ones(100));

# this branch
julia> @time FeatureTransforms.apply(df, lc);
  0.300694 seconds (1.21 M allocations: 63.126 MiB, 3.88% gc time)
  0.000259 seconds (751 allocations: 199.766 KiB)

# main
julia> @time FeatureTransforms.apply(df, lc);
  0.204158 seconds (708.95 k allocations: 34.896 MiB, 6.52% gc time)
  0.003044 seconds (50.54 k allocations: 1.179 MiB)

julia> lc = LinearCombination(ones(5));

# this branch
julia> @time FeatureTransforms.apply(df, lc, cols=[:x1, :x2, :x3, :x4, :x5]);
  0.005228 seconds (4.03 k allocations: 266.237 KiB)
  0.000144 seconds (87 allocations: 19.406 KiB)

# main
julia> @time FeatureTransforms.apply(df, lc, cols=[:x1, :x2, :x3, :x4, :x5]);
  0.155136 seconds (394.81 k allocations: 19.945 MiB, 7.72% gc time)
  0.000211 seconds (2.94 k allocations: 102.297 KiB)